# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "College Lesson"
    },
...
    "paths": {
        "/api/entry": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Вход/Регистрация",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "Username"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password"
                                    },
                                    "team": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "required": [
                                    "username",
                                    "password",
                                    "team"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешный вход"
                                        },
                                        "redirect": {
                                            "type": "string",
                                            "example": "/lesson"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный ввод данных"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь не найден"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exit": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Выход",
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешный выход"
                                        },
                                        "redirect": {
                                            "type": "string",
                                            "example": "/"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка выхода"
                                        },
                                        "redirect": {
                                            "type": "string",
                                            "example": "/"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lesson": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Обновление этапа урока",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "action": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "required": [
                                    "action"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Этап урока успешно обновился"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный ввод данных"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/questions": {
            "get": {
                "tags": [
                    "API"
                ],
                "summary": "Получение вопросов 1 теста",
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questions": {
                                            "type": "object",
                                            "example": {
                                                "id": "Questions_1",
                                                "question": "Какое число?",
                                                "answers": [
                                                    "0",
                                                    "2",
                                                    "3",
                                                    "5",
                                                    "6"
                                                ],
                                                "socer": 5
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/answer": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Отправка ответа на 1 тест",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "answer": {
                                        "example": [
                                            {
                                                "id": "Questions_1",
                                                "answer": "2"
                                            },
                                            {
                                                "id": "Questions_2",
                                                "answer": "5"
                                            }
                                        ]
                                    }
                                },
                                "required": [
                                    "answer"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ответ получен"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Нельзя отправить ответ"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь не найден"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/time": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Старт таймера (lesson/only/team)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "change_time": {
                                        "type": "string",
                                        "example": "lesson"
                                    }
                                },
                                "required": [
                                    "change_time"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешное действие со временем"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный ввод данных"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/time/edit": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Редактирование таймера (lesson/only/team)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "change_time": {
                                        "type": "string",
                                        "example": "lesson"
                                    },
                                    "new_time": {
                                        "type": "string",
                                        "example": "02:00"
                                    }
                                },
                                "required": [
                                    "change_time",
                                    "new_time"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешное действие со временем"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный ввод данных"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/clear": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Очистка данных",
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Данные очищены"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clear/lesson": {
            "post": {
                "tags": [
                    "API"
                ],
                "summary": "Очистка этапа урока",
                "responses": {
                    "202": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Данные очищены"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка в очистке этапа урока"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}